x64 char*/guint cast fixes from chpe.
git-svn-id: http://svn.gnome.org/svn/at-spi/trunk@821 e2bd861d-eb25-0410-b326-f6ed22b6b98c

+++ atk-bridge/bridge.c
@@ -21,6 +21,8 @@
  * Boston, MA 02111-1307, USA.
  */
 
+#include "config.h"
+
 #include <string.h>
 #include <stdio.h>
 #include <unistd.h>
@@ -311,7 +313,7 @@ spi_display_name (void)
 		canonical_display_name = g_strdup (display_env);
 		display_p = strrchr (canonical_display_name, ':');
 		screen_p = strrchr (canonical_display_name, '.');
-		if (screen_p && display_p && ((guint) screen_p > (guint) display_p))
+		if (screen_p && display_p && (screen_p > display_p))
 		{
 		    *screen_p = '\0';
 		}
